--- a/WalletWasabi/Crypto/Groups/Generators.cs
+++ b/WalletWasabi/Crypto/Groups/Generators.cs
@@ -1,5 +1,7 @@
 using NBitcoin.Secp256k1;
+using System.Linq;
 using System.Text;
+using WalletWasabi.Helpers;
 
 namespace WalletWasabi.Crypto.Groups
 {
@@ -97,6 +99,34 @@ namespace WalletWasabi.Crypto.Groups
 			while (!GE.TryCreateXQuad(new FE(buffer), out ge));
 
 			return new GroupElement(ge);
+		}
+
+
+		private static GroupElement[]? _negateGh2i;
+		public static GroupElement[] NegateGh2i 
+		{
+			get
+			{
+				if (_negateGh2i is null)
+				{
+					var negatedGh = Gh.Negate();
+					_negateGh2i = BinaryVector.Select(b => b * negatedGh).ToArray();
+				}
+				return _negateGh2i;
+			}
+		}
+
+		private static Scalar[]? _binaryVector; 
+		public static Scalar[] BinaryVector 
+		{
+			get
+			{
+				if (_binaryVector is null)
+				{
+					_binaryVector = Enumerable.Range(0, Constants.RangeProofWidth).Select(i => Scalar.Zero.CAddBit((uint)i, 1)).ToArray();
+				}
+				return _binaryVector;
+			}
 		}
 	}
 }