--- a/WalletWasabi/Crypto/Groups/GroupElement.cs
+++ b/WalletWasabi/Crypto/Groups/GroupElement.cs
@@ -45,7 +45,7 @@ namespace WalletWasabi.Crypto.Groups
 
 		private GEJ Gej { get; }
 		private Lazy<GE> LazyGe { get; }
-		private GE Ge => LazyGe.Value;
+		internal GE Ge => LazyGe.Value;
 		private bool IsGeCreated => LazyGe.IsValueCreated;
 
 		public bool IsInfinity => Gej.IsInfinity;