--- a/WalletWasabi/WabiSabi/Crypto/WabiSabiClient.cs
+++ b/WalletWasabi/WabiSabi/Crypto/WabiSabiClient.cs
@@ -144,7 +144,7 @@ namespace WalletWasabi.WabiSabi.Crypto
 				var scalarAmount = new Scalar((ulong)amount.Satoshi);
 
 				var randomness = RandomNumberGenerator.GetScalar(allowZero: false);
-				var ma = scalarAmount * Generators.Gg + randomness * Generators.Gh;
+				var ma = ProofSystem.PedersenCommitment(scalarAmount, randomness);
 
 				var (rangeKnowledge, bitCommitments) = ProofSystem.RangeProofKnowledge(scalarAmount, randomness, Constants.RangeProofWidth, RandomNumberGenerator);
 				knowledgeToProve.Add(rangeKnowledge);