--- a/WalletWasabi/WalletWasabi.csproj
+++ b/WalletWasabi/WalletWasabi.csproj
@@ -28,8 +28,8 @@
 	<ItemGroup>
 		<PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="5.0.0" />
 		<PackageReference Include="Microsoft.Win32.SystemEvents" Version="5.0.0" />
-		<PackageReference Include="NBitcoin" Version="5.0.65" />
-		<PackageReference Include="NBitcoin.Secp256k1" Version="1.0.8" />
+		<PackageReference Include="NBitcoin" Version="5.0.71" />
+		<PackageReference Include="NBitcoin.Secp256k1" Version="1.0.10" />
 	</ItemGroup>
 
 	<ItemGroup>