Improves the WabiSabi credentials range proofs performance.

### Before this PR:

```
|  Method |           Job |       Runtime |    Toolchain | k |    Mean |    Error |   StdDev | Rank |
|-------- |-------------- |-------------- |------------- |-- |--------:|---------:|---------:|-----:|
| Process |    Job-OHFFZZ | .NET Core 5.0 | netcoreapp50 | 2 | 8.758 s | 0.2226 s | 0.6494 s |    1 |
```

### After this PR

```
|  Method |           Job |       Runtime |    Toolchain | k |    Mean |    Error |   StdDev | Rank |
|-------- |-------------- |-------------- |------------- |-- |--------:|---------:|---------:|-----:|
| Process |    Job-WTFSTS | .NET Core 5.0 | netcoreapp50 | 2 | 1.042 s | 0.0080 s | 0.0071 s |    1 |
```