@lontivero 2021-01-22
I am not sure this is enough and there are no many things that can be done to improve it further other than changing the rangeproof algorithm. It would be good if we can fine tune the code to gain some extra performance (ideally subsecond execution time). For example, I can see that the batch algorithm used is peppenger but we could try with strauss.