@kiminuo 2021-01-22
@lontivero Can you share how you measure the performance?