@lontivero 2021-01-22
Some extra gains.
```
|  Method |           Job |       Runtime |    Toolchain | k |     Mean |    Error |  StdDev | Rank |
|-------- |-------------- |-------------- |------------- |-- |---------:|---------:|--------:|-----:|
| Process |    Job-NASTIZ | .NET Core 5.0 | netcoreapp50 | 2 | 982.5 ms | 11.57 ms | 9.66 ms |    1 |
```