@lontivero 2021-01-22
@kiminuo you can use this project: https://github.com/zkSNACKs/WasabiBenchmark (i need to update it). Do the following:

* clone it
* remove the WasabiWallet module and create a symbolic link to you WasabiWallet folder
* follow the instructions in the README file.