@lontivero 2021-01-22
I think we can merge this version as it is because it can help the devs implementing the protocol to be a bit more productive because UT doesn't need to run forever now. We can continue discussing changes in a new PR.