@nothingmuch 2021-01-22
Agreed, I think we've reached a Pareto frontier, all the net wins have been added, and any additional changes (apart from CodeNitpicker) can be addressed in subsequent PRs depending on whether the tradeoff seems appropriate.