@lontivero 2021-01-22
Codefactor is wrong, it seems it doesn't understand the c# 9 switch statement syntax.